ml-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
setuptools>=68.0.0
